add a validated style guide
Too many sloppy errors were made due to the lack of an automated style
guide.  This commit adds google's style guide checker and incorporates
it into the build.

This is likely to be un-done if the work ever goes into ApiTrace, but
at least there will be fewer errors in the code when it is reformatted
to the ApiTrace guidelines.