Track retraced program ids instead of traced ids
Mesa deterministically chooses ids in glCreateProgram, which allowed the
program id in a Call object to be used in place of the id generated
during the Call's retrace.

When retracing a linux file on other drivers, the ids will not match.
Frameretrace needs to look up the retraced ids in order to override and
restore shaders.