OpenFile: Display valid frame count if selected frame is too high
A typical user error is providing a frame number that is higher than the
last frame in the trace.  In this case, FrameRetrace crashed.

Instead, the retrace process now sends a fatal error to the host, and
stops processing requests.  The host displays the message, and stops
processing enqueued requests.  The UI quits after the user has accepted
the message dialogue.